third_party/blink/web_tests/external/wpt/css/cssom/font-variant-shorthand-serialization-expected.txt

This is a testharness.js-based test.
[FAIL] font: menu serialization
  assert_equals: font-variant was set expected "" but got "normal"
Harness: the test ran to completion.
